requirements.txt
pytz==2023.3
requests==2.31.0
schedule==1.2.0
python-telegram-bot==20.7
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
flask==3.0.0
sqlalchemy==2.0.23
pyrebase4==4.9.0